+++ allauth/socialaccount/providers/meetup/tests.py
@@ -1,11 +1,12 @@
-from allauth.socialaccount.providers import registry
-from allauth.socialaccount.tests import create_oauth2_tests
-from allauth.tests import MockedResponse
+from allauth.socialaccount.tests import OAuth2TestsMixin
+from allauth.tests import MockedResponse, TestCase
 
 from .provider import MeetupProvider
 
 
-class MeetupTests(create_oauth2_tests(registry.by_id(MeetupProvider.id))):
+class MeetupTests(OAuth2TestsMixin, TestCase):
+    provider_id = MeetupProvider.id
+
     def get_mocked_response(self):
         return MockedResponse(
             200,