+++ allauth/socialaccount/providers/untappd/tests.py
@@ -1,11 +1,12 @@
-from allauth.socialaccount.providers import registry
-from allauth.socialaccount.tests import create_oauth2_tests
-from allauth.tests import MockedResponse
+from allauth.socialaccount.tests import OAuth2TestsMixin
+from allauth.tests import MockedResponse, TestCase
 
 from .provider import UntappdProvider
 
 
-class UntappdTests(create_oauth2_tests(registry.by_id(UntappdProvider.id))):
+class UntappdTests(OAuth2TestsMixin, TestCase):
+    provider_id = UntappdProvider.id
+
     def get_login_response_json(self, with_refresh_token=True):
         return """
             {